chore: moving server to app & add env
