+++ server/config/custom-environment-variables.ts
@@ -10,5 +10,6 @@ export default {
     accessTokenTtl: 'ACCESS_TOKEN_TTL',
     refreshTokenTtl: 'REFRESH_TOKEN_TTL',
     saltWorkerFactor: 'SALT_WORKER_FACTOR',
-    origin: 'ORIGIN'
+    origin: 'ORIGIN',
+    app_env: 'NODE_ENV'
 }
